Show all UI, change icon of weather depend on data from WeatherAPI

+++ app/src/main/java/com/via/weatherapp/models/Sys.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 data class Sys(
     val type: Int,
     val country: String,
-    val sunrise: Int,
-    val sunset: Int
+    val sunrise: Long,
+    val sunset: Long
 ) : Serializable
